+++ sdk/lib/core/iterable.dart
@@ -822,22 +822,3 @@ class _GeneratorIterable<E> extends ListIterable<E> {
   /// Helper function used as default _generator function.
   static int _id(int n) => n;
 }
-
-/// An [Iterator] that allows moving backwards as well as forwards.
-///
-/// Deprecation note: This interface has turned out to not be
-/// valuable as a general reusable interface. There is still only
-/// one class implementing it, [RuneIterator], and at least one other
-/// bidirectional iterator preferred a different name than `movePrevious`
-/// for the other direction.
-/// As such, this interface does not carry its own weight, and will be
-/// removed in a later release.
-@Deprecated("Use the implementing class directly")
-abstract class BidirectionalIterator<E> implements Iterator<E> {
-  /// Move back to the previous element.
-  ///
-  /// Returns true and updates [current] if successful. Returns false
-  /// and updates [current] to an implementation defined state if there is no
-  /// previous element
-  bool movePrevious();
-}