+++ sdk/lib/math/point.dart
@@ -36,7 +36,7 @@ class Point<T extends num> {
   bool operator ==(Object other) =>
       other is Point && x == other.x && y == other.y;
 
-  int get hashCode => SystemHash.hash2(x.hashCode, y.hashCode);
+  int get hashCode => SystemHash.hash2(x.hashCode, y.hashCode, 0);
 
   /// Add [other] to `this`, as if both points were vectors.
   ///